+++ route.go
@@ -9,6 +9,7 @@ func route() {
 	http.HandleFunc("/home", model.Home)
 	http.HandleFunc("/pbt", model.Pbt)
 	http.HandleFunc("/reproto", model.RebuildPbt)
+	http.HandleFunc("/api/xiaodu", model.FromXiaodu)
 	http.HandleFunc("/json", model.ProtocolJSON)
 	http.HandleFunc("/statistic", model.HomeStatistic)
 	http.HandleFunc("/statistics", model.AdvancedStatistic)
