+++ template/list.htm
@@ -89,6 +89,14 @@
         <a href="/reproto">Unmarshal/Marshal Protobuf</a>
         </td>
     </tr>
+    <tr>
+        <td class="docs">
+        <p>Protobuf API x-du</p>
+        </td>
+        <td class="code leading">
+        <a href="/api/xiaodu">Protobuf API sample</a>
+        </td>
+    </tr>
 </table>
